slack integration completed with basic message (#8)
* slack integration completed with basic message

* minor change in timeout

* added condition in aftereach

---------

+++ src/tests/ui/E01.test.ts
@@ -1,4 +1,5 @@
 import { test } from '@fixtures/allPageFixture';
+import type { TestInfo } from '@playwright/test';
 import jsonUtility from '@utils/common/jsonUtility';
 import type { Shopper } from 'types/env.types';
 
@@ -12,7 +13,7 @@ test.beforeEach('Login to app', async ({ pages }) => {
     await pages.factory.loginActions.loginAs(userData);
 });
 
-test('Buy a product via Add to cart', async ({ pages }) => {
+test('Buy a product with cod', { tag: ['@smoke', '@regression', '@cod'] }, async ({ pages }) => {
     await pages.factory.headerActions.searchForItem(productName);
     await pages.factory.productDescriptionActions.clickProduct(productName);
     await pages.factory.productDescriptionActions.addToCart();
@@ -20,11 +21,16 @@ test('Buy a product via Add to cart', async ({ pages }) => {
     await pages.factory.cartActions.clickBuyNow();
     await pages.factory.addressActions.selectAddress(addressId);
     const orderId: string = await pages.factory.paymentActions.codPayment('COD');
+    await pages.apiHelper.changeOrderStatus('DELIVERED', orderId);
     await pages.factory.headerActions.clickMenu();
     await pages.factory.accountSettingActions.selectMenu('My Orders');
-    await pages.factory.myOrdersActions.cancelOrderByOrderId(orderId);
+    await pages.factory.myOrdersActions.validateOrderStatus('DELIVERED', orderId);
+    await pages.factory.headerActions.clickMenu();
+    await pages.factory.accountSettingActions.selectMenu('Logout');
 });
 
-test.afterEach('Cleanup Cart', async ({ pages }) => {
-    await pages.apiHelper.deleteAllProductFromCart();
+test.afterEach('Cleanup Cart', async ({ pages }, testInfo: TestInfo) => {
+    if (testInfo.status !== 'passed' && testInfo.status !== 'timedOut') {
+        await pages.apiHelper.deleteAllProductFromCart();
+    }
 });
